- Populate orders list.

+++ woodroid/src/main/java/me/gilo/woodroid/models/ShippingLine.kt
@@ -8,5 +8,5 @@ class ShippingLine {
     var id: String? = null
     @SerializedName("method_title")
     var methodTitle: String? = null
-    var total: Int = 0
+    var total: Double = 0.0
 }
